+++ README.md
@@ -16,7 +16,7 @@ Features:
 ### Install
 You can install this tool via pip:
 ```
-pip install pyhttpcap
+pip install pycapture
 ```
 
 ### Analysis pcap file
