+++ pycapture/constant.py
