+++ pycapture/processor.py
