+++ pycapture/reader.py
