+++ pycapture/textutils.py
